requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
ocrmypdf==16.5.0
img2pdf==0.5.1
pymupdf4llm==0.0.17
pymupdf==1.24.10
pillow==10.4.0
python-multipart==0.0.9
python-magic==0.4.27
openai
tiktoken
tenacity
pydantic
pydantic-settings>=2.0.0
qdrant-client
psycopg2
celery
spacy
nltk
pyvis
requests
pdf2image
pytesseract>=0.3.10
openpyxl>=3.1.0
xlrd>=2.0.1
tabulate>=0.9.0
redis>=5.0.0  # Added for Celery Redis transport